Add extra resetPagination flag
Sometimes resetting the pagination is not desired, this commit introduces the option to choose when to reset the pagination or not.

This also fixes an issue when triggering a `custom callback` in page `1 + n` with `n>0`. When that `callback` is triggered `componentWillReceiveProps` will be called and the user will be sent back to `page number 1` without any noticed or information about what happened.